Added a new TCP experimental backend for DTM (no MPI, just TCP). To use it, user currently need to follow these three steps : 1) Uncomment the corresponding line in dtm/taskmanager.py ('from commManagerTCP import DtmCommThread')
2) Create a text file with one worker IP address per line (if you want to put more than one worker on the same host, just add the same address more than one time), for instance :
127.0.0.1
127.0.0.1
192.168.0.14
will start two workers on the local machine, and one on the 192.168.0.14

3) As Python cannot starts himself a process on a remote computer, we have to use SSH, only for the launching process. You can start those process manually by giving them two arguments : the
name of the hosts file and their id, for instance :
$ python yourFile.py -f hostFile -i 2

As this procedure is not too easy or practicle, a small Python script was written to make this job automatically (dtm/tests/launchProcess.py). You have to pass as arguments to this script the
name of the program you want to execute and the name of the hosts file, like :
$ python launchProcess.py yourFile.py hostFile


This backend was tested on a single computer (with all the connections routed to localhost), but using it in a 'real' network environment should not create problems -- except the increased
communication latency. Take note that this backend is still highly experimental.


Also made some changes in taskmanager, like removing the singleton class, as it seems to change nothing at all for the imports, but completely mess up Sphinx for the upcoming documentation.