Created new branch for DTM development (dtm-0.3-dev).
Added new features to DTM :
- waitAll(list) and waitAny(list), which wait for the completion of all or any of the tasks passed in args (when the user does not specify a list, default is to use all the current async tasks)
- testAll(list) and testAny(list), which returns whether all or any of the tasks have completed (the same behavior applied if the user does not supply a list)
Take note that waitAll does not return anything (it just guarantees that all the async tasks have completed), while testAll return a bool (True if all the tasks have completed).
waitAny returns a handle (AsyncResult object) specifying which task have completed, or None if no async task have been launched yet.
testAny behaves the same, but always return the last async task completed (the most recent one), even if called multiple times, and return None if no async task is finished.

Those behaviors are subject to be reviewed and changed if needed.

Also implemented a imap_unordered() function (since the only reason it was not in the API was the lack of a waitAny option), again the behavior of this method has to be discussed.

Some minor changes and bug fixes in other parts of DTM.

--HG--
branch : dtm-0.3-dev