Modified import of numpy in creator.
PyPy latest release (1.6.0) includes a beta module for numpy.
However, the new module doesn't include array, and ndarray, therefore
when running a program using the creator with PyPy, the import doesn't
crashed, but the access to numpy.ndarray is impossible, and therefore
a crash still happen. This should fix this problem until PyPy
implements ndarray.