Corrected a bug a cxUniformPartialyMatched.
In the for loop, there was a missing pair of parenthesis after
the call to random.random. So instead of comparing a random number
to the crossover probability, the operator was comparing the address
of the random.random function, which was always greater then 1, so
the condition was always False, and the operator was never applied.