Corrected a difference between documentation and code in CMA-ES.
The documentation on lambda_ and mu_ computation mentionned the
usage of the function 'floor' when default values were computed.
In the case of lambda_, the actual function used was int.
In the case of mu_, no function were used to truncate the result.
The function now used by both lambda_ and mu_ computation is int
and this is properly described in the documentation.