Fixed an error in section 3.3 of the tutorials.
mutGaussian returns a tuple, not a single individual, however the
tuple return by mutGaussian was assigned to a variable ind2, which
was then compared to a second individual mutant. Since ind2 is a
tuple and mutant is an individual, the return value when checking
if both individual are the same was always false.