Tuned up sortFastND, making it up to 5 times faster.
We now exploit the fact that most of the time some individuals
share the same fitness values. We can therefore sort less elements
by using the unique fitnesses values instead of individuals. This
alters the order of the individuals in the front but not the content
of the front, therefore the algorithm has the same behavior, but
algorithms using selNSGA2 won't produce the exact same results since
their results generally depends on the individual order in the
population.

proof of concept : ga_evosn now run 5x times faster.

--HG--
branch : 0.8-dev