Introduces a new implementation of tree for GP.
The base.Tree class was bloated and inefficient which
contradicts most of DEAP principles. We therefore put
an end to it, and started from scratch.

The GP tree is now a simple list where the node can
be found in depth-first order. The arity of the
primitives is used to delimitate the subtrees.

The STGP and normal GP operators have also been merged,
which simplifies the API without impacting the performance.

The new representation can provide a speedup up to 3x-4x
on some problems, mostly when the evaluation times used to
be smaller than the crossover-mutation-copy times.

--HG--
branch : dev