Added a fast version of the Ant problem, with a C version of the ant simulator. The module can be compiled with the script buildAntSimFast.py. It offers the same interface as the Python AntSimulator class and a fast definition of the progn function.
The only current flaw is that the if_food_ahead function must be enclosed in a partial function in order to work, which can be done with a script like :
def ifFoodAheadPartial(out1, out2):
    return partial(ant.if_food_ahead, out1, out2)

pset.addPrimitive(ifFoodAheadPartial, 2)

Currently, a speedup of about 7x may be obtained by using this fast version. Note that as the simulators exhibits some different behaviors, the results obtained might not be
the same (this C version is based upon the OpenBeagle Ant example). I will look furthur to eliminate those differencies.

--HG--
branch : dev