Modified the DTM loadbalancer to remove useless complexity. The previous probability based model has some major flaws, and should have been updated. However, as Scoop will soon replace DTM, a complete rewrite of the load balancer would have been a waste of time.
Nevertheless, for those who are currently using DTM, this update should increase performance, especially if you are running across a cluster (> 16 workers).
The distribution model is now more naive : a worker spawn equally only the tasks that he created (it uses a round-robin over a randomized workers list).
While this may not be the perfect load balancing in some complex situations (i.e. nested map and async calls), it should avoid most of the critical
performance issues of the previous version (like using only 10 workers when running on 128 cores...).

Please note that this fix is not intended to be published in a new version of DEAP, as DTM will be soon deprecated. It's only goal is to provide a quick way
for existing users to fix perfomance issues (the load balancer is the only part modified, the DTM API and behaviors remain the same).

This fix was tested over a small cluster (32 cores), with Python 2.7.

--HG--
branch : dev