Modify varOr to clone when reproducing.
At first, we didn't clone the individual when doing
a simple reproduction since the individual was not
modified, therefore its fitness should stay the same.
However, in the NSGA2 algorithm we assign a crowding-distance
to each individual. As we speak, the crowding-distance is
not unique for each genotype, but unique for each individual,
therefore if two individual in the population actually are
the same object, the will have the same crowding-distance,
which is not conform with the rest of the algorithm.

For now the only case I could figure out was with NSGA-II,
but it is now impossible that two individuals created using
the reproduction might eventually have different attributes
even if they share the same genotype, therefore we should
clone when reproducing.

--HG--
branch : dev