Correct a bug in sortingnetwork addConnector method.
The method was checking for conflict on the previous
level only. However, the connector should be added
on the last level that does not present any conflict.
The method now loops on the level in reversed order
as long as there are no conflicts, then add the
connector to the last level with no conflicts. If
no level is found, a new level is added.

--HG--
branch : dev