Add a Symbol class in gp to represent expression symbol.
The representation of terminals has to be done using the repr
function. However, for symbols, which are represented by string,
the usage of repr adds a pair of quote around the symbol string
representation i.e.: "'x'" instead of 'x', which causes problem
when evaluating an expression. To counter that, we add a new
class that inherits from Terminal, but return the symbol directly
instead of calling repr.

Also now, repr is called instead of str in __repr__ and toString
functions of the Terminal class. This corrects a bug when using
objects like numpy arrays as terminal expressions.

--HG--
branch : dev