Backout/replace most of changes introduced in edd82b68c692.
Instead of creating a Symbolc class, the Terminal can now
be defined as symbolic. Symbolic terminals are : function
names, variables and constants.

Example :
  pset.addTerminal(math.pi, 'pi')

When stringifying the expression, instead of having 3.14159
in the expression string, we get 'pi'. Also terminals can
now be any kind of objects and this does not impact runtime
when a name is set. For example, we can now add a terminal
that is a numpy array :
  pset.addTerminal(numpy.array(range(5)), 'vector1')

--HG--
branch : dev