In creator, replace check for attribute "__call__" by isinstance(obj, type).
This now allows member function to be added to a class created
with the creator, while maintaining the ability to initialize
attributes with new instances of class. On the other hand, it is
no longer possible to initialize an attribute with the return of
a function. This capability had very little value and was actually
never used nor documented.

--HG--
branch : dev