creator: remove the call to object.__init__ when the base class does not override it.
When the base class does not override the object.__init__ method,
a warning exception is raised when it is called with *args and **kargs.
In 25c91c091fff, we therefore check if the method was different of
object.__init__ to determine if the arguments and the keywords should
be passed or not the it. As in fact, the object __init__ method does
nothing at all, therefore calling it is a strict waste of time. From
now on, if the base class does not override the method, it is simply
not called.

--HG--
branch : dev