Rename argument *method* by *function* in Toolbox.register, and fix a bug when registering class.
What truly is registered in the toolbox are functions and not methods. Although,
they can be method, the more general term is function.

The try/except for attribute error on the dictionary was replaced by a conditional
statement, that now also checks if the function argument is a class, in which case
we do not want to copy its dictionary.

--HG--
branch : dev