Apply most of modifications discussed in Issue 17 on Fitness domination.
This patch add a method dominates and a method __hash__ to the Fitness.
The functions using isDominated are fixed and now use dominates.
The function sortFastND is renamed sortNondominated in prevision of the
integration of a faster nondominated sort function. The documentation of
sortNondominated and assignCrowdingDist is improved.

--HG--
branch : dev