Deprecate and remove lambdify, stringify and evaluate, introducing compile.
gp.compile is simply a merge of lambdify and evaluate. If the
primitive set has one or more argument, compile build a lambda
function, while if the number of argument is zero, it simply
returns the result produced by evaluating the tree.

compile appears to be better choice of word than evaluate and
lambdify, while reducing the number of functions to build an
expression from a tree to a single one.

--HG--
branch : dev