GP: code cleaning + solution for ephemeral constants.
Call to the creator in addEphemeralConstants have been
replaced by an explicit call to type. The new classes
are now stored in the gp namespace instead of the
creator namespace. This is cleaner (I think...) and it
maintains the ability to pickle PrimitiveTree as long
as the primitive set is defined in the main module
global scope. It might be the advanced hour, but I
think it might be it, the solution I have been looking
for, paving the way for the golden master of 1.0.0.

--HG--
branch : dev