GP: fix regression with pypy introduced in 0dd7d52c002f.
PyPy __builtins__ module is not iterable like its CPython
counterpart. It is therefore not possible to check if a
symbol is part of builtins by simply doing 'toto' in __builtins__,
we have to refer explictly to the module dictionnary, like
'toto' in __builtins__.__dict__. This is also compatible
with CPython.

--HG--
branch : dev