Fix the implementation of EDA update method in eda/fctmin.py.
The algorithm was converging, but the implementation was not
conformed to the algorithm presented by the paper Teytaud 2009.
The sigma was originally interpreted as a vector while it is
presented as a vector in the paper.

Also added is the reference to article and the class have been
renamed to represent the algorithm instead of generic EDA.

--HG--
branch : dev