Change the generate default type in gp to pset.ret.
This fixes a problem when a STGP user would forget
to set the type_ argument, the generate function
would request primitive of the type 'object'. The
type_ argument is only to accomodate GP mutations.
It is from now on set to None by default, and
replace by pset.ret if the no other type has been
specified.

--HG--
branch : dev