Fix benchmarks.shekel definition (implementation + doc).
The original description wrongfully asserted that the
matrix A gives the location of the local optima. The
exact position of the optima does not correspond to the
coordinates provided by A because the function is a
sum of multiple terms that influence each other.

The positions in A can however be used as a good starting
point for doing a numerical optimization, as the optima
are close to these positions.