Replace Ephemeral by MetaEphemeral.
The usage of metaclass allow us to stop
registering Ephemeral classes in the gp
module global scope. We also register the
__reduce__ function of MetaEphemeral so
Ephemeral could easily be pickled, as long
as the ephemeral function is picklable
(i.e.: no lambda function). The metaclass
also has a cache based on class id to
avoid creating new object for the same
ephemeral when unpickling.