Add metaclass MetaFitness in base.
The metaclass verifies that weights are provided
at class creation and that they are a sequence.
We should probably add a verification that weights
are actually a sequence of numbers.

This commit also makes DEAP base Fitness class
a mono-objective maximization fitness. This
simplifies MetaFitness check as it would fail
otherwise when defining base.Fitness and it
also means that user won't have to define a
fitness for basic problems.

This is just the beginning as this class could be
used to avoid having to provides a sequence for
mono-objective fitness. It could also be used to
defined an alternative dominance strategy to Pareto
dominance.