Fix dtlz7 benchmark problem implementation
dtlz7 function used to raise ``TypeError: 'float' object has no attribute
'__getitem__'`` on each call.

Problem rose, because the same variable name ``ind`` was used both to
store coordinates of the individual and as a temporal variable in a loop.
Temporal variable used to override the original one used to store coordinates
of the individual. Problem fixed by using different name for the temporal variable.