setup: fix git invocation for exlinks, add override
The 'tree' would be bytes, so we'd end up with an url like
https://github.com/DEAP/deap/blob/b'0123456'/examples/.

While at it, allow the user to override the detect hash. This is particularly
useful when there are local modifications, and HEAD hash doesn't yet exist on
github, and it's nicer to link to some other commit, or even 'master'.

Also, sphinx 6 enforces the second item in the extlinks tuple (the title)
to have exactly one '%s'.