Fixed variable name bug
There is a bug in which zero individuals may be returned from
SPEA-2 selection.  This change addresses that.

The variable `k` is used both as a parameter and as an index in
loops.  This is a bug that appears to have been introduced in 2023;
the original loop variable was `l`.  I have replaced the index
variable with `m`, which avoids problematic `l`, maintains the old
convention that index variables be single letters between 'i' and
'n', and does not overwrite the user-provided parameter.